refactor: 공개 API 규칙 적용 및 파일 구조 리팩토링
