+++ src/main.tsx
@@ -2,7 +2,7 @@ import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import App from '@app/App.tsx';
 import { worker } from '@shared/api/mocks/browser.ts';
-import { config } from '@/shared/utils/config.ts';
+import { config } from '@/shared/config/appConfig';
 import './main.css';
 
 if (config.useMock) {
