+++ src/shared/config/index.ts
@@ -1,0 +1,4 @@
+export * from './appConfig';
+export * from './routes';
+export * from './theme';
+export * from './typography';
