+++ src/shared/ui/dev-panner/RoutingPanel.tsx
@@ -1,4 +1,4 @@
-import { ROUTE_LIST } from '@/shared/utils/routes';
+import { ROUTE_LIST } from '@/shared/config';
 import { useNavigate } from 'react-router-dom';
 
 export function RoutingPanel() {
